Golem
Attack3

+++ 3D BaseFramework/Src/Application/GameObject/UI/Enemy/HP/Enemy_HP.h
@@ -33,7 +33,7 @@ private:
 	Math::Rectangle            m_rect[HP::num];
 	int                        m_MaxHP = 0;
 	int                        m_beforeHP = 0;
-	long                       m_MaxWidth = 0.0l;
+	float                      m_MaxWidth = 0.0f;
 	int                        m_DownTime = 0;
 	float                      m_alpha = 1.0f;
 };
